Added depends_on attribute
A Homebrew formula allows to specify one or multiple formulas that are
needed in order to build the current one.

This is done by the 'depends_on' attribute: https://docs.brew.sh/Formula-Cookbook#specifying-other-formulae-as-dependencies

In this pull request I've added this attribute.

I would be happy to have it merge as I need a dependency in my case ;-)

Thank you